% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/sumR.R
\name{sumRRcpp}
\alias{sumRRcpp}
\alias{sumR}
\title{Addition of two numbers}
\usage{
sumRRcpp(x, y)

sumR(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The addition of \code{x} and \code{y}.
}
\description{
This package contains several functions to compute the addition
of two numbers.
}
\details{
\code{sumRRcpp} is an implementation using the Rcpp interface.
}
\examples{
sumRRcpp(1, 2)

sumR(1, 2)

}
